refactor: changing models to schemas
